+++ test/TerraformDirectoryPluginTest.groovy
@@ -1,17 +1,16 @@
 import static org.hamcrest.Matchers.containsString
 import static org.hamcrest.Matchers.hasItem
 import static org.hamcrest.Matchers.instanceOf
-import static org.junit.Assert.assertThat
+import static org.hamcrest.MatcherAssert.assertThat
 
-import org.junit.After
-import org.junit.Test
-import org.junit.runner.RunWith
-import de.bechte.junit.runners.context.HierarchicalContextRunner
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
 
-@RunWith(HierarchicalContextRunner.class)
 class TerraformDirectoryPluginTest {
+    @Nested
     public class Init {
-        @After
+        @AfterEach
         void resetPlugins() {
             TerraformInitCommand.resetPlugins()
             TerraformValidateCommand.resetPlugins()
@@ -52,12 +51,14 @@ class TerraformDirectoryPluginTest {
         }
     }
 
+    @Nested
     public class Apply {
-        @After
+        @AfterEach
         void resetJenkinsEnv() {
             TerraformDirectoryPlugin.withDirectory("./terraform/")
         }
 
+        @Nested
         public class WithDirectoryProvided {
             @Test
             void addsDirectoryToTerraformInit() {
@@ -108,6 +109,7 @@ class TerraformDirectoryPluginTest {
             }
         }
 
+        @Nested
         public class WithoutDirectoryProvided {
             @Test
             void addsDirectoryToTerraformInit() {