+++ test/TerraformLandscapePluginTest.groovy
@@ -1,24 +1,23 @@
 import static org.hamcrest.Matchers.containsString
 import static org.hamcrest.Matchers.hasItem
 import static org.hamcrest.Matchers.instanceOf
-import static org.junit.Assert.assertThat
+import static org.hamcrest.MatcherAssert.assertThat
 import static org.mockito.Mockito.mock;
 
-import org.junit.Test
-import org.junit.After
-import org.junit.Before
-import org.junit.runner.RunWith
-import de.bechte.junit.runners.context.HierarchicalContextRunner
+import org.junit.jupiter.api.AfterEach
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
 
-@RunWith(HierarchicalContextRunner.class)
 class TerraformLandscapePluginTest {
-    @Before
+    @BeforeEach
     public void resetJenkinsEnv() {
         Jenkinsfile.instance = mock(Jenkinsfile.class)
     }
 
+    @Nested
     public class Init {
-        @After
+        @AfterEach
         void resetPlugins() {
             TerraformPlanCommand.resetPlugins()
         }
@@ -32,8 +31,8 @@ class TerraformLandscapePluginTest {
         }
     }
 
+    @Nested
     public class Apply {
-
         @Test
         void addsLandscapeArgumentToTerraformPlan() {
             TerraformLandscapePlugin plugin = new TerraformLandscapePlugin()